인기영화 더보기 페이징 처리

+++ app/src/main/java/org/techtown/diffuser/activity/detailpage/DetailViewModel.kt
@@ -15,7 +15,7 @@ class DetailViewModel @Inject constructor(
     savedStateHandle: SavedStateHandle,
     private val repository: Repository
 ) : ViewModel() {
-    val movieId = savedStateHandle.get<Int>("movie_id") ?: 0
+    val movieId = savedStateHandle.get<Long>("movie_id") ?: 0
 
     private val _items: MutableLiveData<List<ItemModel>> = MutableLiveData()
     val items: LiveData<List<ItemModel>> = _items
